@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using AutoMapper;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
@@ -16,6 +15,7 @@ using AspNetCoreRateLimit;
 using Microsoft.AspNetCore.Http;
 using reCAPTCHA.AspNetCore;
 using SendGrid;
+using Microsoft.Extensions.Hosting;
 
 [assembly: ApiConventionType(typeof(DefaultApiConventions))]
 namespace QuizyfyAPI
@@ -24,7 +24,6 @@ namespace QuizyfyAPI
     {
         public IConfiguration Configuration { get; }
         private SwaggerOptions SwaggerOptions;
-
 
         public Startup(IConfiguration configuration)
         {
@@ -66,7 +65,7 @@ namespace QuizyfyAPI
             services.AddSingleton<IIpPolicyStore, MemoryCacheIpPolicyStore>();
             services.AddSingleton<IRateLimitCounterStore, MemoryCacheRateLimitCounterStore>();
 
-            services.ConfigureMvcForApi(AppOptions);
+            services.ConfigureControllersForApi(AppOptions);
 
             services.ConfigureValidationErrorResponse();
 
@@ -86,11 +85,11 @@ namespace QuizyfyAPI
 
             services.AddTransient<IRecaptchaService, RecaptchaService>();
             services.AddPwnedPasswordHttpClient();
-            services.AddSingleton<ISendGridClient, SendGridClient>(factory => new SendGridClient(SendGridOptions));
+            _ = services.AddSingleton<ISendGridClient, SendGridClient>(_ => new SendGridClient(SendGridOptions));
             services.AddSingleton<ISendGridService, SendGridService>();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             app.UseIpRateLimiting();
 
@@ -105,12 +104,14 @@ namespace QuizyfyAPI
             }
 
             app.UseHttpsRedirection();
+            app.UseStaticFiles();
+
+            app.UseRouting();
+            app.UseCors("default");
 
             app.UseMiddleware<ExceptionMiddleware>();
-
-            app.UseCors(x => x.AllowAnyOrigin().AllowAnyMethod().AllowAnyHeader());
-
             app.UseAuthentication();
+            app.UseAuthorization();
 
             app.UseSwagger();
 
@@ -119,9 +120,8 @@ namespace QuizyfyAPI
                 setupAction.SwaggerEndpoint(SwaggerOptions.UIEndpoint, SwaggerOptions.Title);
                 setupAction.RoutePrefix = SwaggerOptions.RoutePrefix;
             });
-            app.UseStaticFiles();
 
-            app.UseMvc();
+            app.UseEndpoints(endpoints => endpoints.MapControllers());
         }
     }
 }
