@@ -3,12 +3,12 @@
 namespace QuizyfyAPI.Contracts.Responses
 {
     /// <summary>
-    /// A choice with text(actual answer) and isRight bool. Used for displaying questions and DTO.
+    /// Request used to provide data for email verification.
     /// </summary>
     public class VerifyEmailRequest
     {
         /// <summary>
-        /// Token.
+        /// Email verification token generated by guid.
         /// </summary>
         [Required]
         public string Token { get; set; }
