@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace QuizyfyAPI.Data
 {
@@ -10,6 +8,7 @@ namespace QuizyfyAPI.Data
     {
         [Key]
         public int Id { get; set; }
+
         public int QuizId { get; set; }
         public string Text { get; set; }
         public virtual ICollection<Choice> Choices { get; set; }
