@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace QuizyfyAPI.Domain
+﻿namespace QuizyfyAPI.Domain
 {
     public class DetailedResult : BasicResult
     {
