@@ -15,7 +15,6 @@ using System.Reflection;
 using System.IO;
 using Microsoft.OpenApi.Models;
 using QuizyfyAPI.Options;
-
 namespace QuizyfyAPI
 {
     public static class ServiceExtensions
@@ -29,6 +28,7 @@ namespace QuizyfyAPI
                 options.AssumeDefaultVersionWhenUnspecified = swaggerOptions.SupplyDefaultVersion;
             });
         }
+
         public static void ConfigureDbContext(this IServiceCollection services, AppOptions appOptions)
         {
             services.AddDbContextPool<QuizDbContext>(
@@ -36,6 +36,7 @@ namespace QuizyfyAPI
                     appOptions.ConnectionString
                 ));
         }
+
         public static void ConfigureJWTAuth(this IServiceCollection services, JwtOptions jwtOptions)
         {
             // configure jwt authentication
@@ -79,21 +80,23 @@ namespace QuizyfyAPI
                 x.TokenValidationParameters = tokenValidationParameters;
             });
         }
-        public static void ConfigureMvcForApi(this IServiceCollection services, AppOptions appOptions)
+
+        public static void ConfigureControllersForApi(this IServiceCollection services, AppOptions appOptions)
         {
-            services.AddMvc(setupAction =>
+            services.AddControllers(setupAction =>
             {
                 setupAction.Filters.Add(new ProducesResponseTypeAttribute(StatusCodes.Status500InternalServerError));
 
                 setupAction.ReturnHttpNotAcceptable = appOptions.ReturnHttpNotAcceptable;
 
-                var jsonFormatter = setupAction.OutputFormatters.OfType<JsonOutputFormatter>().FirstOrDefault();
-                if (jsonFormatter != null && jsonFormatter.SupportedMediaTypes.Contains("text/json"))
+                var jsonFormatter = setupAction.OutputFormatters.OfType<SystemTextJsonOutputFormatter>().FirstOrDefault();
+                if (jsonFormatter?.SupportedMediaTypes.Contains("text/json") == true)
                 {
                     jsonFormatter.SupportedMediaTypes.Remove("text/json");
                 }
-            }).SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            }).SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
         }
+
         public static void ConfigureSwagger(this IServiceCollection services, SwaggerOptions swaggerOptions)
         {
             services.AddSwaggerGen(setupAction =>
@@ -140,6 +143,7 @@ namespace QuizyfyAPI
                 setupAction.IncludeXmlComments(xmlCommentsPath);
             });
         }
+
         public static void ConfigureValidationErrorResponse(this IServiceCollection services)
         {
             services.Configure<ApiBehaviorOptions>(options =>
